refactor: break out helper methods into separate file
