+++ decrypt_test.go
@@ -1,8 +1,9 @@
 package avtool
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func Test_Decrypt_V11(t *testing.T) {
